+++ R/callme.R
@@ -15,15 +15,14 @@
 #'        e.g. \code{"-L/opt/homebrew/lib -lzstd"} to include the homebrew 
 #'        libraries in the linker search path and to link to the \code{zstd}
 #'        library installed there. 
+#' @param verbose Should the output of the compiler be echoed to the R console?
+#'        Default: FALSE
 #'        
 #' @export
 #' 
-#' @return A reference to the dynamic library which was loaded into R with 
-#'         \code{dyn.load()} i.e. a 'DLLInfo' object.  This has been setup 
-#'         such that when the returned object goes out of scope, the dynamic 
-#'         library is unloaded i.e. the C function will no longer be accessible.
-#'         Thus the user needs to keep a reference to this returned object for 
-#'         as long as they need access to the C function.
+#' @return A named list of R functions which wrap the calls to the equivalent
+#'         C functions.  When this returned object is garbage collected, the
+#'         generated library will be unloaded.
 #'         
 #' @examples
 #' \dontrun{
@@ -39,16 +38,16 @@
 #' # when \code{'dll'} gets garbage collected.
 #' dll <- callme(code)
 #' 
-#' # Call the function
+#' # Manually call the function
 #' .Call("calc_", 1, 2.5)
 #' 
-#' # Advanced: Get info about the C function / library.
-#' dll$calc_
+#' # Use the auto-generated wrapper function
+#' dll$calc_(1, 2.5)
 #' }
 #' 
 #'
 #~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-callme <- function(code, cpp_flags = NULL, ld_flags = NULL) {
+callme <- function(code, cpp_flags = NULL, ld_flags = NULL, verbose = FALSE) {
   
   #~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
   # Sanity check code
@@ -68,7 +67,7 @@ callme <- function(code, cpp_flags = NULL, ld_flags = NULL) {
   tmp_dir   <- tempfile(pattern = paste0("callme_", datestamp, "_"))
   dir.create(tmp_dir, showWarnings = FALSE, recursive = TRUE)
   
-  tmp_file  <- tempfile(tmpdir = tmp_dir)
+  tmp_file  <- file.path(tmp_dir, paste0("callme_", datestamp))
   c_file    <- paste0(tmp_file, ".c")
   dll_file  <- paste0(tmp_file, .Platform$dynlib.ext)
   
@@ -100,14 +99,42 @@ callme <- function(code, cpp_flags = NULL, ld_flags = NULL) {
   }
   
   #~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+  # Setup stdout/stderr
+  #~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+  if (isTRUE(verbose)) {
+    stdout = ""  # echo to R console
+    stderr = ""  # echo to R console
+  } else {
+    stdout = "stdout"  # divert to file
+    stderr = "stderr"  # divert to file
+  }
+  
+  #~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
   # Compile a shared library.
   # LDFLAGs can be included at the end of a "R CMD SHLIB" call to add
   # extra library search paths and link to libraries.
   #~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-  system2(
+  ret <- system2(
     command = paste0(R.home(component = "bin"), "/R"), 
-    args    = paste("CMD SHLIB", basename(c_file), ld_flags)
+    args    = paste("CMD SHLIB", basename(c_file), ld_flags),
+    stdout  = stdout,
+    stderr  = stdout
   )
+  
+  #~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+  # If there's an error, force show the output even if verbose = FALSE
+  #~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+  if (ret != 0) {
+    warning("Error during compilation")
+    if (!verbose) {
+      if (file.exists('stdout'))
+        cat(readLines("stdout"), sep = "\n")
+      if (file.exists('stderr')) 
+        cat(readLines("stderr"), sep = "\n")
+    }
+    return(NULL)
+  }
+  
   
   #~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
   # Load the DLL
@@ -139,10 +166,113 @@ callme <- function(code, cpp_flags = NULL, ld_flags = NULL) {
   #    * the finalizer() function will be called
   #         * which unloads the actual dll 
   #~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+  res <- list()
   dll[['env']] <- new.env()
   reg.finalizer(dll[['env']], finalizer, onexit = TRUE)
+  res$raw_dll <- dll
+  
+  #~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+  # Automatically generate some wrapper functions
+  #~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+  funcs <- create_wrapper_functions(code)
+  for (fname in names(funcs)) {
+    res[[fname]] <- funcs[[fname]]
+  }
   
   
-  dll
+  res
 }
 
+
+
+#~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+#' Create \code{.Call()} function wrappers for C functions
+#' 
+#' @param code C code as single string.  This does not recurse into source
+#'        or header files referenced in the C code.
+#'        
+#' @return named list of R functions which can call into the library
+#' 
+#' @import stringr
+#' @noRd
+#~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+create_wrapper_functions <- function(code) {
+  
+  decls <- extract_function_declarations(code)
+  funcs <- lapply(decls, create_wrapper_function)
+  
+  unlist(funcs, recursive = FALSE)
+}
+
+
+#~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+#' Extract function declarations which are consistent with .Call() semantics
+#' 
+#' @param code C code as single string.  This does not recurse into source
+#'        or header files referenced in the C code.
+#' @noRd
+#~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+extract_function_declarations <- function(code) {
+  # Find all function declarsions which match: "SEXP func_name(...)"
+  decls <- stringr::str_extract_all(code, stringr::regex("^\\s*SEXP\\s+[a-zA-Z0-9_]+\\s*\\(.*?\\)", multiline = TRUE, dotall=TRUE))[[1]]
+  
+  # tidy function declarations for next step
+  decls <- stringr::str_trim(decls)
+  decls <- stringr::str_replace_all(decls, "\\s+", " ")
+  
+  decls
+}
+
+
+
+#~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+#' Create a wrapper functon for the given C declaration of a .Call compatible 
+#' function
+#' 
+#' @param decl C declaration. E.g. \code{"SEXP two_(SEXP vara, SEXP varb)"}
+#' 
+#' @return anonymous R function to \code{.Call} the dll
+#' 
+#' @import stringr
+#' @noRd
+#~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+create_wrapper_function <- function(decl) {
+  
+  #~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+  # Extract the function name
+  #~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+  func <- stringr::str_match(decl, "SEXP\\s+([a-zA-Z0-9_]+)")
+  stopifnot(nrow(func) == 1)
+  func_name <- func[1, 2]
+  
+  #~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+  # Extract the argument names
+  #~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+  args <- stringr::str_match(decl, "\\((.*?)\\)")
+  stopifnot(nrow(args) == 1)
+  args <- args[1, 2]
+  args <- stringr::str_split(args, ",")[[1]]
+  args <- stringr::str_replace_all(args, "SEXP", "")
+  args <- stringr::str_trim(args)
+  if (length(args) == 1 && args == "") {
+    # Setting args to NULL here makes the function creation easier.
+    args <- NULL
+  }
+  
+  #~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+  # Create the function as a string
+  #~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+  func_str <- sprintf(
+    "function(%s) .Call(%s)", 
+    paste(args, collapse = ", "),
+    paste(c(dQuote(func_name, q=FALSE), args), collapse = ", ")
+  )
+  
+  #~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+  # Parse function into R code and put in a named list
+  #~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+  res <- list(eval(parse(text = func_str)))
+  names(res) <- func_name
+  
+  res
+}